Add files via upload
I changed the definitions of the I_\xi functions and added the J_\xi functions so that the definitions were identical to those from the latest version of the work 	arXiv:2305.07771. These changes do not affect the effects of the software - the results are identical to the results of the previous version.